sdl2: change default color depth
The default color depth does not work on NXP i.MX 8X and i.MX 8M Plus.
The reason is unknown - either some issue on libsdl2 or the graphics
driver.

Simply changing the default allows the game to run properly. It is not a
proper or a real fix though, since it does not solve the underlying
issue.

Learn more on https://github.com/Warzone2100/warzone2100/pull/2561